 More schema attributes added
Explored the possibility of group certain terms in cases of repeated suggestions for simplification and improved usability- no changes due to variations in use across the different findings.

Additional schema attributes were added to describe the behavior of the schema elements better and, by doing so, improve clinical usability.

The schema was validated and converted using HED online tools.